GIX-1187: Raise error in queryAndUpdate
